backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.1
pandas==2.1.4
numpy==1.26.2
httpx==0.25.2
python-multipart==0.0.6